requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
gunicorn==21.2.0
python-dotenv==1.0.0
pymongo==4.5.0
Flask-WTF==1.1.1
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
python-dateutil==2.8.2
dnspython==2.4.2
Pillow==10.0.0 